+++ Assets/OxGFrame/APICenter/Scripts/Runtime/APIBase.cs
@@ -1,6 +1,0 @@
-﻿namespace OxGFrame.APICenter
-{
-    public abstract class APIBase
-    {
-    }
-}