+++ Assets/OxGFrame/AgencyCenter/Scripts/Runtime/APICenter/APIBase.cs
@@ -1,0 +1,6 @@
+﻿namespace OxGFrame.AgencyCenter.APICenter
+{
+    public abstract class APIBase
+    {
+    }
+}