+++ Assets/OxGFrame/AgencyCenter/Scripts/Runtime/EventCenter/EventBase.cs
@@ -1,0 +1,11 @@
+﻿using Cysharp.Threading.Tasks;
+
+namespace OxGFrame.AgencyCenter.EventCenter
+{
+    public abstract class EventBase
+    {
+        public abstract UniTaskVoid HandleEvent();
+
+        protected abstract void Release();
+    }
+}